Introduce (=/=) to mark equations that are known to be not proved
This is mostly useful to document what one would like to be able to
prove eventually.

Also, search for proof tasks under arguments, so that you can write

    foo x = lhs x === rhs x

instead of

    foo = (\x -> lhs x) === (\x -> rhs x)

But careful: BangPatterns for example will make the plugin not find the
(===) any more. (Maybe it should traverse the tree and complain if it
finds it anywhere else.)